mavlink: move ATTITUDE_QUATERNION to streams hpp
